FIND - tests added
Tests FIND with an empty string and a non-existent word. The non-existent word is the same as in exceptiontest.fth. This should be defined in future in a configuration file. This is defined using ANS Forth CHAR instead of Forth 2012 character input 'x' as some systems may delay implementation of this feature.